Ajustada estrutura de pastas de app
